// src/types/project.types.ts
export interface IProject {
  id: string;
  projectDuration?: string;
  projectName: string;
  githubClientLink: string;
  githubServerLink: string;
  liveSiteLink: string;
  technologies: string;
  featureOne: string;
  featureTwo: string;
  featureThree: string;
  createdAt: string;
  updatedAt: string;
}